fix: harden api diff and async providers
