+++ scripts/generate-component-api-docs.ts
@@ -4,6 +4,15 @@ import { execFile } from "node:child_process";
 import { promisify } from "node:util";
 import { fileURLToPath } from "node:url";
 import ts from "typescript";
+import {
+  ambiguousPublicPropNames,
+  componentEventPayloads,
+  componentPublicEventDescriptions,
+  componentPublicPropDescriptions,
+  publicEventDescriptions,
+  publicEventPayloads,
+  sharedPublicPropDescriptions,
+} from "./api-doc-metadata.js";
 
 const execFileAsync = promisify(execFile);
 
@@ -43,492 +52,6 @@ type ComponentMeta = {
   entrypoint: string;
   props: PropMeta[];
   events: EventMeta[];
-};
-
-const ambiguousPublicPropNames = new Set(["closeOnSelect", "items", "rows", "selectable", "value"]);
-
-const sharedPublicPropDescriptions: Record<string, string> = {
-  x: "Left position in terminal cells.",
-  y: "Top position in terminal cells.",
-  w: "Width in terminal cells.",
-  h: "Height in terminal cells.",
-  zIndex: "Render and event ordering within the current plane.",
-  modelValue: "Controlled component value.",
-  tone: "Semantic color tone.",
-  style: "Base terminal cell style override.",
-  title: "Optional title text.",
-  titleStyle: "Style override for title text.",
-  label: "Visible label text.",
-  name: "Field name used by form context.",
-  labelStyle: "Style override for label text.",
-  disabled: "Disables pointer and keyboard activation.",
-  disabledStyle: "Style used for disabled content.",
-  selectedStyle: "Style used for selected rows or nodes.",
-  highlightStyle: "Style used for the highlighted row or match.",
-  highlightMatchStyle: "Style used for highlighted text while the row is active.",
-  matchStyle: "Style used for matched text.",
-  emptyStyle: "Style used when rendering an empty state.",
-  emptyText: "Text rendered when there are no rows or items.",
-  clear: "Clears the component rectangle before drawing content.",
-  wrap: "Wraps text to the available cell width.",
-  border: "Draws a border around the component.",
-  borderStyle: "Style override for border cells.",
-  padding: "Inner padding in terminal cells.",
-  scrollX: "Horizontal content offset in terminal cells.",
-  scrollY: "Vertical content offset in terminal cells.",
-  focusable: "Adds the component to keyboard focus navigation.",
-  selectionScrollBy: "Scroll callback used while a pointer selection reaches the viewport edge.",
-  autoFocus: "Requests focus when the component becomes visible.",
-  cols: "Terminal column count.",
-  widthProvider: "Cell width provider used by the terminal buffer.",
-  defaultStyle: "Default terminal cell style for descendants.",
-  theme: "Theme token overrides for component defaults.",
-  autoResize: "Resizes the terminal from the host element when enabled.",
-  minCols: "Minimum column count used by auto resize.",
-  minRows: "Minimum row count used by auto resize.",
-  recordEvents: "Optional event recorder callback.",
-  inputPlugins: "Input plugins provided to descendant text inputs.",
-  pathPickerProvider: "Path provider injected into descendant path pickers.",
-  linkOpener: "Host link opener used by components with host-owned activation.",
-  debugIme: "Enables IME debugging output.",
-  debugTrace: "Enables runtime trace logging.",
-  domRendererOptions: "DOM renderer options used by TerminalProvider.",
-  clipboard: "Clipboard implementation used for terminal selection copy.",
-  selection: "Terminal cell selection configuration.",
-  href: "Link target to render and activate.",
-  hoverStyle: "Style applied while the pointer hovers the link.",
-  focusStyle: "Style applied while the link has keyboard focus.",
-  activeStyle: "Style used for the active item or enabled state.",
-  visited: "Marks the link as already visited for styling.",
-  openMode: "Link activation mode.",
-  activationKeys: "Keyboard keys that activate the link.",
-  modifierClick: "Pointer modifier required for click activation.",
-  linkStyle: "Style applied to detected link segments.",
-  protocols: "URL protocols accepted by linkification.",
-  allowRelative: "Allows relative hrefs in detected link segments.",
-  maxUrlLength: "Maximum detected URL length.",
-  columns: "Table column definitions.",
-  minWidth:
-    "Minimum column width in terminal cells; treated as a soft minimum when space is constrained.",
-  maxWidth: "Maximum column width in terminal cells.",
-  flex: "Relative width share for auto-sized table columns.",
-  contentStyle: "Style override for dialog or popover content cells.",
-  rowKey: "Row key field or resolver.",
-  selectedRowKey: "Controlled selected row key.",
-  selectedRowKeys: "Controlled selected row keys for multi-select tables.",
-  scrollTop: "Controlled top row offset.",
-  header: "Shows the table header when enabled.",
-  headerStyle: "Style override for table header cells.",
-  headerFocusable: "Makes header cells keyboard focusable.",
-  rowFocusable: "Makes body rows keyboard focusable.",
-  sortable: "Enables sortable column header interactions.",
-  sortBy: "Controlled sorted column key.",
-  sortDirection: "Controlled sort direction.",
-  manualSort: "Disables built-in sorting while keeping sort events controlled by the host.",
-  sorter: "Custom row comparison function.",
-  filter: "Controlled filter query.",
-  filterable: "Enables built-in row filtering.",
-  manualFilter: "Disables built-in filtering while keeping filter state host-owned.",
-  filterPredicate: "Custom row filter predicate.",
-  selectionMode: "Row selection mode.",
-  selectedIndex: "Controlled active item index.",
-  query: "Search query used by filtering or async providers.",
-  itemsProvider: "Async command provider called with the current query.",
-  matcher: "Custom command matcher.",
-  filterStrategy: "Built-in command matching strategy.",
-  itemVersion: "External version key for item changes that keep array identity stable.",
-  multiple: "Enables multi-select mode.",
-  multipleEmit:
-    "Payload shape used by multi-select change and confirm events; the label emit mode emits option labels.",
-  closeOnBlur: "Emits close when focus leaves the component.",
-  initialQuery: "Query used when the command palette opens.",
-  showRowDetails: "Shows command detail text next to labels.",
-  placeholder: "Placeholder text shown when the input is empty.",
-  placeholderWhenFocused: "Placeholder text used while the input has focus.",
-  noMatchesText: "Text rendered when filtering returns no commands.",
-  loadingText: "Text rendered while async commands are loading.",
-  errorText: "Text rendered when async commands fail to load.",
-  debounce: "Delay before calling an async provider, in milliseconds.",
-  minQueryLength: "Minimum query length before async loading runs.",
-  maxVisibleItems: "Maximum number of command rows rendered at once.",
-  resetQueryOnClose: "Resets the query when the palette closes.",
-  hint: "Footer hint text.",
-  hintStyle: "Style override for hint text.",
-  chromeStyle: "Style override for command palette chrome.",
-  inputStyle: "Style override for the embedded input.",
-  listStyle: "Style override for list rows.",
-  bodyStyle: "Style override for dialog body cells.",
-  dividerStyle: "Style override for dividers.",
-  detailStyle: "Style override for detail text.",
-  backdrop: "Renders a backdrop behind the dialog.",
-  backdropStyle: "Style override for backdrop cells.",
-  placement: "Dialog placement within the current layout.",
-  offsetX: "Horizontal placement offset in cells.",
-  offsetY: "Vertical placement offset in cells.",
-  closeOnBackdrop: "Closes the dialog when the backdrop is clicked.",
-  closeOnEsc: "Closes the dialog on Escape.",
-  closeOnConfirm: "Closes the dialog after a footer button confirms.",
-  teleport: "Mounts the dialog into the overlay runtime plane.",
-  tabMode: "Keyboard Tab behavior inside the dialog.",
-  buttons: "Dialog footer buttons.",
-  checkedStyle: "Style used when the checkbox is checked.",
-  options: "Options rendered by the control.",
-  optionProvider: "Async option provider called with the current query.",
-  valueMode: "Model value shape emitted by the select v-model.",
-  activeIndex: "Controlled active option index.",
-  searchable:
-    "Emits query updates from typed characters; local options are not filtered automatically.",
-  typeahead: "Enables keyboard typeahead navigation.",
-  loading: "Shows the loading row.",
-  maxVisible: "Maximum number of option rows rendered at once.",
-  min: "Minimum numeric value.",
-  max: "Maximum numeric value.",
-  step: "Keyboard increment step.",
-  help: "Help text rendered below the field.",
-  helpStyle: "Style override for help text.",
-  error: "Error text rendered below the field.",
-  errorStyle: "Style override for error text.",
-  required: "Marks the field label as required.",
-  suggestions: "Autocomplete suggestions.",
-  suggestionProvider: "Async suggestion provider called with the current input value.",
-  open: "Controlled suggestion popup visibility.",
-  highlightedIndex: "Controlled highlighted suggestion index.",
-  minChars: "Minimum input length before suggestions are shown or loaded.",
-  filterLocal: "Filters provided suggestions against the input value.",
-  suggestionStyle: "Style override for suggestion rows.",
-  activeSuggestionStyle: "Style override for the active suggestion row.",
-  nodes: "Tree nodes.",
-  expandedIds: "Controlled expanded tree node ids.",
-  selectedId: "Controlled selected tree node id.",
-  indent: "Indent width per tree depth.",
-  selectableParents: "Allows expandable parent tree nodes to be selected from their label.",
-  cursorToEndOnExternalUpdate: "Moves the cursor to the end after external model updates.",
-  cursorToEndOnFirstFocus: "Moves the cursor to the end on first focus.",
-  cursorBlink: "Enables cursor blink rendering.",
-  cursorShape: "Cursor glyph shape.",
-  blinkInterval: "Cursor blink interval in milliseconds.",
-  promptSuggestions: "Prompt popup suggestions.",
-  promptTrigger: "Prompt popup trigger character.",
-  promptTriggers: "Prompt popup trigger characters.",
-  promptMaxItems: "Maximum prompt popup rows.",
-  promptAlign: "Prompt popup alignment.",
-  promptSelectedStyle: "Style override for the active prompt suggestion.",
-  promptPopupStyle: "Style override for the prompt popup body.",
-  promptPopupBorderStyle: "Style override for the prompt popup border.",
-  promptPopupMatchStyle: "Style override for prompt match highlights.",
-  skillTrigger: "Trigger used for skill suggestions.",
-  skillSuggestions: "Skill suggestions shown in the prompt popup.",
-  skillHighlightStyle: "Style override for highlighted skill chips.",
-  mentionTrigger: "Trigger used for path or mention suggestions.",
-  mentionWorkspace: "Workspace root used by mention providers.",
-  mentionMode: "Mention provider mode.",
-  mentionShowHidden: "Includes hidden paths in mention suggestions.",
-  mentionSuggestions: "Mention suggestions supplied by the host.",
-  mentionMaxItems: "Maximum mention rows.",
-  mentionChipStyle: "Style override for mention chips.",
-  multilineChipStyle: "Style override for multiline chips.",
-  dedupeMentions: "Removes duplicate mention entries.",
-  collectMentions: "Collects mention values from committed input.",
-  mentions: "Controlled collected mention values.",
-  collapseMultiline: "Collapses multiline pasted text into chips.",
-  multilineTexts: "Controlled multiline chip text values.",
-  secret: "Masks input text when enabled.",
-  maskChar: "Character used to mask secret input.",
-  submitOnEnter: "Emits change on Enter.",
-  clearOnEscape: "Clears the input on Escape.",
-  plugins: "Input plugins attached to this input.",
-  pasteImageHandler: "Host handler for pasted images.",
-  filePasteHandler: "Host handler for pasted files.",
-};
-
-function pickSharedPublicPropDescriptions(...names: string[]): Record<string, string> {
-  return Object.fromEntries(names.map((name) => [name, sharedPublicPropDescriptions[name]!]));
-}
-
-const componentPublicPropDescriptions: Record<string, Record<string, string>> = {
-  TAutocompleteInput: {
-    ...pickSharedPublicPropDescriptions(
-      "modelValue",
-      "suggestions",
-      "suggestionProvider",
-      "open",
-      "highlightedIndex",
-      "debounce",
-      "minChars",
-      "filterLocal",
-    ),
-    closeOnSelect: "Closes suggestions after a suggestion is selected.",
-  },
-  TBadge: {
-    ...pickSharedPublicPropDescriptions("tone"),
-    value: "Text or scalar value rendered by the badge.",
-  },
-  TCheckbox: {
-    ...pickSharedPublicPropDescriptions("modelValue"),
-  },
-  TCode: {
-    value: "Code text rendered inside the code block.",
-  },
-  TCommandPalette: {
-    ...pickSharedPublicPropDescriptions(
-      "modelValue",
-      "query",
-      "initialQuery",
-      "itemsProvider",
-      "matcher",
-      "filterStrategy",
-      "selectedIndex",
-      "showRowDetails",
-      "noMatchesText",
-      "hint",
-      "debounce",
-      "minQueryLength",
-      "maxVisibleItems",
-      "resetQueryOnClose",
-    ),
-    closeOnSelect: "Closes the command palette after a command is selected.",
-    items: "Command items rendered and filtered by the palette.",
-  },
-  TDataTable: {
-    ...pickSharedPublicPropDescriptions(
-      "columns",
-      "rowKey",
-      "selectedRowKey",
-      "selectedRowKeys",
-      "scrollTop",
-      "sortDirection",
-      "sortable",
-      "manualSort",
-      "sorter",
-      "filter",
-      "filterable",
-      "manualFilter",
-      "filterPredicate",
-      "selectionMode",
-    ),
-    selectable: "Enables row selection.",
-  },
-  TDialog: {
-    ...pickSharedPublicPropDescriptions(
-      "modelValue",
-      "placement",
-      "offsetX",
-      "offsetY",
-      "backdrop",
-      "closeOnBackdrop",
-      "closeOnEsc",
-      "teleport",
-      "tabMode",
-      "buttons",
-      "closeOnConfirm",
-    ),
-  },
-  TFormField: {
-    ...pickSharedPublicPropDescriptions("name", "help", "error", "required"),
-  },
-  TInput: {
-    ...pickSharedPublicPropDescriptions(
-      "modelValue",
-      "cursorToEndOnExternalUpdate",
-      "cursorToEndOnFirstFocus",
-      "cursorBlink",
-      "cursorShape",
-      "blinkInterval",
-      "promptSuggestions",
-      "promptTrigger",
-      "promptTriggers",
-      "promptMaxItems",
-      "promptAlign",
-      "promptSelectedStyle",
-      "promptPopupStyle",
-      "promptPopupBorderStyle",
-      "promptPopupMatchStyle",
-      "skillTrigger",
-      "skillSuggestions",
-      "skillHighlightStyle",
-      "mentionTrigger",
-      "mentionWorkspace",
-      "mentionMode",
-      "mentionShowHidden",
-      "mentionSuggestions",
-      "mentionMaxItems",
-      "mentionChipStyle",
-      "multilineChipStyle",
-      "dedupeMentions",
-      "collectMentions",
-      "mentions",
-      "collapseMultiline",
-      "multilineTexts",
-      "secret",
-      "maskChar",
-      "submitOnEnter",
-      "clearOnEscape",
-      "plugins",
-      "pasteImageHandler",
-      "filePasteHandler",
-    ),
-  },
-  TLink: {
-    ...pickSharedPublicPropDescriptions(
-      "href",
-      "visited",
-      "openMode",
-      "activationKeys",
-      "modifierClick",
-    ),
-  },
-  TLinkifyText: {
-    ...pickSharedPublicPropDescriptions("protocols", "allowRelative", "maxUrlLength"),
-    value: "Text scanned for links and rendered into terminal cells.",
-  },
-  TList: {
-    ...pickSharedPublicPropDescriptions("itemVersion", "modelValue"),
-    items: "List rows rendered by the component.",
-  },
-  TPasswordInput: {
-    ...pickSharedPublicPropDescriptions("modelValue"),
-  },
-  TRadioGroup: {
-    ...pickSharedPublicPropDescriptions("modelValue", "options"),
-  },
-  TSelect: {
-    ...pickSharedPublicPropDescriptions(
-      "options",
-      "optionProvider",
-      "query",
-      "modelValue",
-      "valueMode",
-      "activeIndex",
-      "multiple",
-      "multipleEmit",
-      "searchable",
-      "typeahead",
-      "debounce",
-      "loading",
-      "maxVisible",
-    ),
-  },
-  TSlider: {
-    ...pickSharedPublicPropDescriptions("modelValue", "min", "max", "step"),
-  },
-  TSwitch: {
-    ...pickSharedPublicPropDescriptions("modelValue"),
-  },
-  TTable: {
-    ...pickSharedPublicPropDescriptions(
-      "columns",
-      "rowKey",
-      "selectedRowKey",
-      "selectedRowKeys",
-      "header",
-      "headerFocusable",
-      "rowFocusable",
-    ),
-  },
-  TTag: {
-    ...pickSharedPublicPropDescriptions("tone"),
-  },
-  TerminalProvider: {
-    ...pickSharedPublicPropDescriptions(
-      "cols",
-      "widthProvider",
-      "defaultStyle",
-      "theme",
-      "autoResize",
-      "minCols",
-      "minRows",
-      "recordEvents",
-      "inputPlugins",
-      "pathPickerProvider",
-      "linkOpener",
-      "debugIme",
-      "debugTrace",
-      "domRendererOptions",
-      "clipboard",
-      "selection",
-    ),
-    rows: "Terminal row count.",
-  },
-  TText: {
-    value: "Text content rendered into terminal cells.",
-  },
-  TTree: {
-    ...pickSharedPublicPropDescriptions(
-      "nodes",
-      "expandedIds",
-      "selectedId",
-      "indent",
-      "selectableParents",
-    ),
-  },
-  TView: {
-    selectable: "Controls whether terminal text selection may start inside the view.",
-  },
-};
-
-const publicEventDescriptions: Record<string, string> = {
-  "update:modelValue": "Emitted when the controlled model value changes.",
-  "update:selectedIndex": "Emitted when the controlled active index changes.",
-  "update:highlightedIndex": "Emitted when the active autocomplete suggestion changes.",
-  "update:open": "Emitted when popup visibility changes.",
-  "update:query": "Emitted when the controlled query changes.",
-  "update:activeIndex": "Emitted when the active option index changes.",
-  "update:selectedRowKey": "Emitted when the selected row key changes.",
-  "update:selectedRowKeys": "Emitted when selected row keys change.",
-  "update:scrollTop": "Emitted when the top visible row offset should change.",
-  "update:sortBy": "Emitted when the sorted column key changes.",
-  "update:sortDirection": "Emitted when the sort direction changes.",
-  "update:expandedIds": "Emitted when expanded tree node ids change.",
-  "update:selectedId": "Emitted when the selected tree node id changes.",
-  "update:mentions": "Emitted when collected mentions change.",
-  "update:multilineTexts": "Emitted when multiline chip text values change.",
-  input: "Emitted for input edits.",
-  change: "Emitted when the component commits a value change.",
-  select: "Emitted when the active item is selected.",
-  close: "Emitted when the component requests to close.",
-  focus: "Emitted when the component receives focus.",
-  blur: "Emitted when the component loses focus.",
-  keydown: "Emitted for keydown events.",
-  keyup: "Emitted for keyup events.",
-  click: "Emitted for click events.",
-  dblclick: "Emitted for double-click events.",
-  pointerdown: "Emitted for pointer down events.",
-  pointerup: "Emitted for pointer up events.",
-  pointermove: "Emitted for pointer move events.",
-  pointerenter: "Emitted when the pointer enters the component.",
-  pointerleave: "Emitted when the pointer leaves the component.",
-  wheel: "Emitted for wheel events.",
-  confirm: "Emitted when a focused action is confirmed.",
-  activate: "Emitted when the link is activated.",
-  open: "Emitted when the host opener accepts a link open request.",
-  invalidHref: "Emitted when a link href is rejected by the sanitizer.",
-  scroll: "Emitted when the visible scroll offset changes.",
-  sortChange: "Emitted when table sort state changes.",
-  rowSelect: "Emitted when a data table row is selected.",
-  rowClick: "Emitted when a table row is clicked or confirmed.",
-  rowKeydown: "Emitted when a focused table row receives a keydown event.",
-  headerClick: "Emitted when a table header is clicked or confirmed.",
-  toggle: "Emitted when a tree node expands or collapses.",
-  selectionCopy: "Emitted after terminal selection copy.",
-  mentionClick: "Emitted when a rendered mention chip is clicked.",
-  multilineClick: "Emitted when a multiline chip is clicked.",
-  validationError: "Emitted when input validation rejects a host action.",
-};
-
-const publicEventPayloads: Record<string, string> = {
-  focus: "void",
-  blur: "void",
-  close: "void",
-  pointerenter: "TerminalPointerEvent",
-  pointerleave: "TerminalPointerEvent",
-  pointerdown: "TerminalPointerEvent",
-  pointerup: "TerminalPointerEvent",
-  pointermove: "TerminalPointerEvent",
-  click: "TerminalPointerEvent",
-  dblclick: "TerminalPointerEvent",
-  wheel: "TerminalPointerEvent",
-  keydown: "TerminalKeyboardEvent",
-  keyup: "TerminalKeyboardEvent",
 };
 
 type ApiManifest = {
@@ -1095,32 +618,8 @@ function inferEventPayload(component: ComponentMeta, event: EventMeta): PayloadR
     };
   }
   const byComponent = `${component.name}.${event.name}`;
-  const componentPayloads: Record<string, string> = {
-    "TAutocompleteInput.select": "TAutocompleteSelectPayload",
-    "TCommandPalette.select": "TCommandPaletteSelectPayload",
-    "TDataTable.rowSelect": "TDataTableRowSelectPayload",
-    "TDataTable.sortChange": "TDataTableSortChangePayload",
-    "TTable.rowKeydown": "TTableRowKeydownPayload",
-    "TDialog.confirm": "DialogButton & { index: number }",
-    "TInput.change": "string",
-    "TInput.input": "string",
-    "TInput.mentionClick": "(absPath: string, event: TerminalPointerEvent)",
-    "TInput.multilineClick": "number",
-    "TInput.validationError": "{ reason: string }",
-    "TList.change": "{ index: number; value: string }",
-    "TList.scroll": "number",
-    "TPasswordInput.change": "string",
-    "TPasswordInput.input": "string",
-    "TSelect.change": "string | string[] | number[] | TSelectMultipleChangePayload | null",
-    "TSelect.confirm": "string[] | number[] | TSelectMultipleChangePayload",
-    "TTable.headerClick": "TTableHeaderClickPayload",
-    "TTable.rowClick": "TTableRowClickPayload",
-    "TTree.select": "TTreeSelectPayload",
-    "TTree.toggle": "TTreeTogglePayload",
-    "TerminalProvider.selectionCopy": "TerminalSelectionCopyPayload",
-  };
-  if (componentPayloads[byComponent]) {
-    return { payload: componentPayloads[byComponent], source: "component-default" };
+  if (componentEventPayloads[byComponent]) {
+    return { payload: componentEventPayloads[byComponent], source: "component-default" };
   }
   if (event.name === "change" || event.name === "input") {
     return {
@@ -1134,9 +633,13 @@ function inferEventPayload(component: ComponentMeta, event: EventMeta): PayloadR
   };
 }
 
-function describeEvent(event: EventMeta): DescriptionResult {
+function describeEvent(component: ComponentMeta, event: EventMeta): DescriptionResult {
   if (event.description) {
     return { description: event.description, source: event.descriptionSource };
+  }
+  const componentDescription = componentPublicEventDescriptions[`${component.name}.${event.name}`];
+  if (componentDescription) {
+    return { description: componentDescription, source: "component-default" };
   }
   const baseName = eventBaseName(event.name);
   if (event.name.endsWith("Capture")) {
@@ -1175,7 +678,7 @@ function fillPublicDocDefaults(component: ComponentMeta): ComponentMeta {
     }),
     events: component.events.map((event) => {
       const payload = inferEventPayload(component, event);
-      const description = describeEvent(event);
+      const description = describeEvent(component, event);
       return {
         ...event,
         payload: payload.payload,
